pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "weather api"
authors = ["Tanja <tanja.grozdani@loka.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
